Improve code accorfing to review.Changed name of activity.Outsource the creation of barcode. Moving activity to package ticket
