+++ app/src/main/java/de/tum/in/tumcampusapp/component/ui/news/KinoDetailsFragment.java
@@ -21,11 +21,10 @@ import com.squareup.picasso.Callback;
 import com.squareup.picasso.Picasso;
 
 import de.tum.in.tumcampusapp.R;
-import de.tum.in.tumcampusapp.api.tumonline.TUMOnlineRequest;
 import de.tum.in.tumcampusapp.component.ui.news.repository.KinoLocalRepository;
 import de.tum.in.tumcampusapp.component.ui.news.repository.KinoRemoteRepository;
 import de.tum.in.tumcampusapp.component.ui.tufilm.model.Kino;
-import de.tum.in.tumcampusapp.component.ui.tufilm.show_ticket;
+import de.tum.in.tumcampusapp.component.ui.ticket.ShowTicketActivity;
 import de.tum.in.tumcampusapp.database.TcaDb;
 import de.tum.in.tumcampusapp.utils.Const;
 import de.tum.in.tumcampusapp.utils.DateUtils;
@@ -171,12 +170,12 @@ public class KinoDetailsFragment extends Fragment {
 
         rootView.addView(headerView);
     }
-    //open show_ticket activity, and transfer current movie data to it.
-    //The String data is just a example, it should be get from correspongding movie which now show on the screen.
-    //time and place of movie  is fixed. Only the movie title and date should be transfered to show_ticket activity
+    //open ShowTicketActivity activity, and transfer current movie data to it.
+    //TODO: The String data is just a example, it should be get from correspongding movie which now show on the screen.
+    //Time and place of movie  is fixed. Only the movie title and date should be transfered to ShowTicketActivity activity
     private void showTicket(){
         String data = "KingsMan 08.05 " + "\n"+ "Filmbegin: 20:00 o'clock "+"\n"+ " 1. Stock, Hörsaal 1200 (Carl-von-Linde-Hörsaal) Arcisstraße 21";
-        Intent intent = new Intent(getActivity().getApplicationContext(), show_ticket.class);
+        Intent intent = new Intent(getActivity().getApplicationContext(), ShowTicketActivity.class);
         intent.putExtra("movie_data", data);
         startActivity(intent);
     }
